* deleted redundant old stuff * PSOWalker (currently theo janssen walker) * fix build
git-svn-id: http://subversion.assembla.com/svn/ncrobot@12 539a5a7e-e5a0-4617-9da5-5f7f46e2b2b9

+++ code/swarm/swarm.cpp
@@ -83,7 +83,7 @@ void Swarm::readInput()
           
   if(friendships==-4)
   {
-    int z=sqrt(populationSize);
+    int z = (int)sqrtf(populationSize);
     if(z*z!=populationSize) z++;
   
     //horizontal edges
